リファクタリング

+++ Editor/App.cpp
@@ -155,10 +155,10 @@ LRESULT CALLBACK App::WndProc(HWND hwnd, UINT message, WPARAM wparam, LPARAM lpa
 				app->editor->OnKeyUp(static_cast<int>(wparam));
 				return 0;
 			case WM_LBUTTONDOWN:
-				app->editor->OnLButtonDown(GET_X_LPARAM(lparam), GET_Y_LPARAM(lparam));
+				app->editor->OnLButtonDown(static_cast<float>(GET_X_LPARAM(lparam)), static_cast<float>(GET_Y_LPARAM(lparam)));
 				return 0;
 			case WM_LBUTTONUP:
-				app->editor->OnLButtonUp(GET_X_LPARAM(lparam), GET_Y_LPARAM(lparam));
+				app->editor->OnLButtonUp(static_cast<float>(GET_X_LPARAM(lparam)), static_cast<float>(GET_Y_LPARAM(lparam)));
 				return 0;
 			case WM_IME_SETCONTEXT:
 				lparam &= ~ISC_SHOWUICOMPOSITIONWINDOW;
